Add support for DESTDIR variable
This allows (among others) packagers to install the files in a temporary
folder to generate a RPM or DEB file. The DESTDIR variable is a well
established convention for this.

This commit also adds the possibility to override the installation PREFIX,
BIN_DIR and DATA_DIR (as some distros would prefer /usr/games for the binary
and /usr/share/games/tbftss for the data).